finish registration setup
