+++ src/components/PostList.css
@@ -19,6 +19,7 @@
   color: white;
   border-radius: 12px;
   cursor: pointer;
+  margin-left: 1.5rem;
 }
 
 .feed-header-action span {
