+++ src/pages/PostDetail.tsx
@@ -1,15 +1,19 @@
 import React from 'react';
-import { useSelector } from 'react-redux';
-import { FaPencil } from 'react-icons/fa6';
+
 import Post from '../components/Post';
 import './PostDetail.css';
 import Nagivation from '../components/Navigation';
 import Header from '../components/Header';
+import { useEffect } from 'react';
+import { useAppDispatch, useAppSelector } from '../store/hooks';
+import { getCurrentUser } from '../store/userSlice';
 
 const PostAnalytics: React.FC = () => {
-  // const { posts } = useSelector((store) => store.post);
+  const dispatch = useAppDispatch();
 
-  // const filteredPosts = posts.filter((post) => post.posts.length !== 0);
+  useEffect(() => {
+    dispatch(getCurrentUser());
+  }, []);
 
   return (
     <div className="post-detail-container">
